Clean all code.
