+++ .styleci.yml
@@ -1,7 +1,7 @@
 preset: psr12
 risky: true
 
-version: 8.1
+version: 8.2
 
 finder:
   exclude:
