+++ changelog.md
@@ -1,2 +1,6 @@
-Change Log
-==========
+Tailwind Asset Change Log
+=========================
+
+## 1.0.0 May 19, 2023
+
+- Initial release.
