+++ package.json
@@ -2,7 +2,7 @@
     "name": "tailwind-asset",
     "version": "1.0.0",
     "dependencies": {
-        "tailwindcss": "^3.2.7"
+        "tailwindcss": "^3.3"
     },
     "license": "mit"
 }
