--- src/TailwindCdnAsset.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yii\Tailwind\Asset;
-
-use Yiisoft\Assets\AssetBundle;
-
-final class TailwindCdnAsset extends AssetBundle
-{
-    public bool $cdn = true;
-    public array $css = ['https://cdn.jsdelivr.net/npm/tailwindcss/dist/tailwind.min.css'];
-}
